+++ WhaleTalk/WhaleTalk/FirebaseModel.swift
@@ -1,0 +1,15 @@
+//
+//  FirebaseModel.swift
+//  WhaleTalk
+//
+//  Created by Ngo Thai on 5/2/16.
+//  Copyright © 2016 TBLStudio. All rights reserved.
+//
+
+import Foundation
+import Firebase
+import CoreData
+
+protocol FirebaseModel {
+    func upload(rootRef: Firebase, context: NSManagedObjectContext)
+}
